+++ packages/presets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @monitor-sdk/presets
 
+## 0.1.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @monitor-sdk/performance@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes